+++ x/btcstaking/keeper/btc_delegations.go
@@ -18,7 +18,11 @@ import (
 // - indexing the given BTC delegation in the BTC delegator store,
 // - saving it under BTC delegation store, and
 // - emit events about this BTC delegation.
-func (k Keeper) AddBTCDelegation(ctx sdk.Context, btcDel *types.BTCDelegation) error {
+func (k Keeper) AddBTCDelegation(
+	ctx sdk.Context,
+	btcDel *types.BTCDelegation,
+	minUnbondingTime uint32,
+) error {
 	if err := btcDel.ValidateBasic(); err != nil {
 		return err
 	}
@@ -75,9 +79,9 @@ func (k Keeper) AddBTCDelegation(ctx sdk.Context, btcDel *types.BTCDelegation) e
 			StakingTxHash: stakingTxHash.String(),
 			NewState:      types.BTCDelegationStatus_UNBONDED,
 		})
-		// NOTE: we should have verified that EndHeight > btcTip.Height + CheckpointFinalizationTimeout
-		wValue := k.btccKeeper.GetParams(ctx).CheckpointFinalizationTimeout
-		k.addPowerDistUpdateEvent(ctx, btcDel.EndHeight-wValue, unbondedEvent)
+
+		// NOTE: we should have verified that EndHeight > btcTip.Height + max(w, min_unbonding_time)
+		k.addPowerDistUpdateEvent(ctx, btcDel.EndHeight-minUnbondingTime, unbondedEvent)
 	}
 
 	return nil
@@ -156,9 +160,9 @@ func (k Keeper) addCovenantSigsToBTCDelegation(
 func (k Keeper) btcUndelegate(
 	ctx sdk.Context,
 	btcDel *types.BTCDelegation,
-	unbondingTxSig *bbn.BIP340Signature,
+	u *types.DelegatorUnbondingInfo,
 ) {
-	btcDel.BtcUndelegation.DelegatorUnbondingSig = unbondingTxSig
+	btcDel.BtcUndelegation.DelegatorUnbondingInfo = u
 	k.setBTCDelegation(ctx, btcDel)
 
 	if !btcDel.HasInclusionProof() {