+++ x/btcstaking/types/query.go
@@ -4,6 +4,18 @@ import (
 	"encoding/hex"
 )
 
+func delegatorUnbondingInfoToResponse(ui *DelegatorUnbondingInfo) *DelegatorUnbondingInfoResponse {
+	var spendStakeTxHex = ""
+
+	if len(ui.SpendStakeTx) > 0 {
+		spendStakeTxHex = hex.EncodeToString(ui.SpendStakeTx)
+	}
+
+	return &DelegatorUnbondingInfoResponse{
+		SpendStakeTxHex: spendStakeTxHex,
+	}
+}
+
 // NewBTCDelegationResponse returns a new delegation response structure.
 func NewBTCDelegationResponse(btcDel *BTCDelegation, status BTCDelegationStatus) (resp *BTCDelegationResponse) {
 	resp = &BTCDelegationResponse{
@@ -44,8 +56,8 @@ func (ud *BTCUndelegation) ToResponse() (resp *BTCUndelegationResponse) {
 		CovenantSlashingSigs:     ud.CovenantSlashingSigs,
 	}
 
-	if ud.DelegatorUnbondingSig != nil {
-		resp.DelegatorUnbondingSigHex = ud.DelegatorUnbondingSig.ToHexStr()
+	if ud.DelegatorUnbondingInfo != nil {
+		resp.DelegatorUnbondingInfoResponse = delegatorUnbondingInfoToResponse(ud.DelegatorUnbondingInfo)
 	}
 	if ud.SlashingTx != nil {
 		resp.SlashingTxHex = ud.SlashingTx.ToHexStr()